Fix the ATString for both controller. This fixes the XBee communiation issue
So I did more digging and change the ATString for both the rotation controller and the shutter controller.
I gave them each their own address, forced some parameters and removed the RO0 command... and low and behold
I see all the commands sent by the rotation controller hello.
Rotation :
ATCE1,ID7734,CH0C,MY0,DH0,DLFFFF,AP0,SM0,WR,CN
So we force the PAN Id and the channel, as well as the rotation controller id to 0 (MY0) and the destination
to the broadcast address. We also set the rotation controller as the coordinator

Shutter :
ATCE0,ID7734,CH0C,MY1,DH0,DLFFFF,AP0,SM0,WR,CN
Same thing but with a CE0  to set tit as an end device, MY1 to set the address to 1.